+++ src/main/kotlin/com/github/atm1020/tuilaunch/services/TuiLauncherSettings.kt
@@ -30,6 +30,7 @@ class TuiLauncherSettings : PersistentStateComponent<TuiLauncherSettings.State> 
         var rememberPromptBoxVisibilityPerApp: Boolean = false,
         var rememberPromptBoxSizePerApp: Boolean = false,
         var focusTuiAfterPromptBoxSend: Boolean = false,
+        var hidePromptBoxAfterSend: Boolean = false,
         var promptBoxCompletionSource: PromptBoxCompletionSource = PromptBoxCompletionSource.JETBRAINS_AI,
         var copilotLanguageServerPath: String = "",
         var copilotPromptHistoryContext: Boolean = true,