+++ src/test/kotlin/com/github/atm1020/tuilaunch/SendPromptBoxActionTest.kt
@@ -35,6 +35,7 @@ class SendPromptBoxActionTest : BasePlatformTestCase() {
             submitPromptOnSend = false
             appendPromptSeparatorOnSend = false
             focusTuiAfterPromptBoxSend = false
+            hidePromptBoxAfterSend = false
         }
         promptBoxFocusRequest = { }
         promptBoxHoldsFocus = { false }